Resolve a minor spelling issue
Resolve the warning message
Warning message:
The vignette title specified in \VignetteIndexEntry{} is different from the title in the YAML metadata. The former is "Adding phylomaps to phylomapr", and the latter is "Adding phylomaps to `phylomapr`". If that is intentional, you may set options(rmarkdown.html_vignette.check_title = FALSE) to suppress this check.